requirements-default.txt
accelerate==1.12.0
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
astor==0.8.1
asttokens==3.0.1
async-lru==2.0.5
attrs==25.4.0
av==16.0.1
babel==2.17.0
beautifulsoup4==4.14.2
blake3==1.0.8
bleach==6.3.0
cachetools==6.2.1
cbor2==5.7.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
cloudpickle==3.1.2
comm==0.2.3
compressed-tensors==0.11.0
contourpy==1.3.3
cupy-cuda12x==13.6.0
cycler==0.12.1
db-dtypes
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
depyf==0.19.0
dill==0.4.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.8.0
einops==0.8.1
email-validator==2.3.0
executing==2.2.1
fastapi==0.121.1
fastapi-cli==0.0.16
fastapi-cloud-cli==0.3.1
fastjsonschema==2.21.2
fastrlock==0.8.3
filelock==3.20.0
fire==0.7.1
fonttools==4.60.1
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.8.0
fsspec==2025.10.0
fuzzywuzzy==0.18.0
gguf==0.17.1
google-cloud-bigquery
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
interegular==0.3.3
importlib-resources
ipykernel==7.1.0
ipython==9.7.0
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
jedi==0.19.2
jinja2==3.1.6
jiter==0.12.0
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-client==8.6.3
jupyter-console==6.6.3
jupyter-core==5.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter-server==2.17.0
jupyter-server-terminals==0.5.3
jupyterlab==4.5.0
jupyterlab-pygments==0.3.0
jupyterlab-server==2.28.0
jupyterlab-widgets==3.0.16
kiwisolver==1.4.9
lark==1.2.2
llguidance==0.7.30
llvmlite==0.44.0
lm-format-enforcer==0.11.3
lmdeploy==0.10.2
markdown-it-py==4.0.0
markupsafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
mistral-common==1.8.5
mistune==3.1.4
mmengine-lite==0.10.7
mpmath==1.3.0
msgpack==1.1.2
msgspec==0.19.0
multidict==6.7.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.5
nibabel
ninja==1.13.0
notebook==7.5.0
notebook-shim==0.2.4
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.580.82
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
openai==2.7.2
openai-harmony==0.0.8
opencv-python-headless==4.12.0.88
outlines-core==0.2.11
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
partial-json-parser==0.2.1.1.post6
peft==0.14.0
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
prometheus-client==0.23.1
prometheus-fastapi-instrumentator==7.1.0
prompt-toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.0
psutil==7.1.3
ptyprocess==0.7.0
pure-eval==0.2.3
py-cpuinfo==9.0.0
pyarrow
pybase64==1.4.2
pycountry==24.6.1
pycparser==2.23
pydantic==2.12.4
pydantic-core==2.41.5
pydantic-extra-types==2.10.6
pygments==2.19.2
pynvml==13.0.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.3
pyzmq==27.1.0
qwen-vl-utils==0.0.14
ray==2.51.1
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.6
rpds-py==0.28.0
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
send2trash==1.8.3
sentencepiece==0.2.1
sentry-sdk==2.43.0
setproctitle==1.3.7
setuptools==79.0.1
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.8
soxr==1.0.0
stack-data==0.6.3
starlette==0.49.3
sympy==1.14.0
termcolor==3.2.0
terminado==0.18.1
threadpoolctl==3.6.0
tiktoken==0.12.0
tinycss2==1.4.0
tokenizers==0.22.1
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.1
transformers-stream-generator==0.0.5
triton==3.4.0
typer==0.20.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
vllm==0.11.0
watchfiles==1.1.1
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
widgetsnbextension==4.0.15
xformers==0.0.32.post1
xgrammar==0.1.25
yapf==0.43.0
yarl==1.22.0
